// src/yeying/api/application/application.ts
export interface SearchApplicationRequest {
  header: MessageHeader | undefined;
  body: SearchApplicationRequestBody | undefined;
}

export interface SearchApplicationRequestBody {
  condition: SearchApplicationCondition | undefined;
  page: RequestPage | undefined;
}

export interface SearchApplicationCondition {
  /** 应用类型【参数可选】 */
  code: ApplicationCodeEnum;
  /** 应用状态【参数可选】 */
  status: ApplicationStatusEnum;
  /** 应用所有者【参数可选】 */
  owner: string;
  /** 应用名称【参数可选】 */
  name: string;
  /** 输入[应用所有者、应用名称]关键字搜索【参数可选】 */
  keyword: string;
}

export interface SearchApplicationResponse {
  header: MessageHeader | undefined;
  body: SearchApplicationResponseBody | undefined;
}

export interface SearchApplicationResponseBody {
  status: ResponseStatus | undefined;
  applications: ApplicationMetadata[];
  page: ResponsePage | undefined;
}

export interface CreateApplicationRequest {
  header: MessageHeader | undefined;
  body: CreateApplicationRequestBody | undefined;
}

export interface CreateApplicationRequestBody {
  application: ApplicationMetadata | undefined;
}

export interface CreateApplicationResponse {
  header: MessageHeader | undefined;
  body: CreateApplicationResponseBody | undefined;
}

export interface CreateApplicationResponseBody {
  status: ResponseStatus | undefined;
  application: ApplicationMetadata | undefined;
}

export interface ApplicationDetailRequest {
  header: MessageHeader | undefined;
  body: ApplicationDetailRequestBody | undefined;
}

export interface ApplicationDetailRequestBody {
  did: string;
  version: number;
}

export interface ApplicationDetailResponse {
  header: MessageHeader | undefined;
  body: ApplicationDetailResponseBody | undefined;
}

export interface ApplicationDetailResponseBody {
  status: ResponseStatus | undefined;
  application: ApplicationMetadata | undefined;
}

export interface DeleteApplicationRequest {
  header: MessageHeader | undefined;
  body: DeleteApplicationRequestBody | undefined;
}

export interface DeleteApplicationRequestBody {
  did: string;
  version: number;
}

export interface DeleteApplicationResponse {
  header: MessageHeader | undefined;
  body: DeleteApplicationResponseBody | undefined;
}
// ...
export const CreateApplicationRequest: MessageFns<CreateApplicationRequest> = {
  encode(message: CreateApplicationRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.header !== undefined) {
      MessageHeader.encode(message.header, writer.uint32(10).fork()).join();
    }
    if (message.body !== undefined) {
      CreateApplicationRequestBody.encode(message.body, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateApplicationRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateApplicationRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.header = MessageHeader.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.body = CreateApplicationRequestBody.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateApplicationRequest {
    return {
      header: isSet(object.header) ? MessageHeader.fromJSON(object.header) : undefined,
      body: isSet(object.body) ? CreateApplicationRequestBody.fromJSON(object.body) : undefined,
    };
  },

  toJSON(message: CreateApplicationRequest): unknown {
    const obj: any = {};
    if (message.header !== undefined) {
      obj.header = MessageHeader.toJSON(message.header);
    }
    if (message.body !== undefined) {
      obj.body = CreateApplicationRequestBody.toJSON(message.body);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateApplicationRequest>, I>>(base?: I): CreateApplicationRequest {
    return CreateApplicationRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateApplicationRequest>, I>>(object: I): CreateApplicationRequest {
    const message = createBaseCreateApplicationRequest();
    message.header = (object.header !== undefined && object.header !== null)
      ? MessageHeader.fromPartial(object.header)
      : undefined;
    message.body = (object.body !== undefined && object.body !== null)
      ? CreateApplicationRequestBody.fromPartial(object.body)
      : undefined;
    return message;
  },
};
// ...
export const ApplicationDetailRequest: MessageFns<ApplicationDetailRequest> = {
  encode(message: ApplicationDetailRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.header !== undefined) {
      MessageHeader.encode(message.header, writer.uint32(10).fork()).join();
    }
    if (message.body !== undefined) {
      ApplicationDetailRequestBody.encode(message.body, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ApplicationDetailRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseApplicationDetailRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.header = MessageHeader.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.body = ApplicationDetailRequestBody.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ApplicationDetailRequest {
    return {
      header: isSet(object.header) ? MessageHeader.fromJSON(object.header) : undefined,
      body: isSet(object.body) ? ApplicationDetailRequestBody.fromJSON(object.body) : undefined,
    };
  },

  toJSON(message: ApplicationDetailRequest): unknown {
    const obj: any = {};
    if (message.header !== undefined) {
      obj.header = MessageHeader.toJSON(message.header);
    }
    if (message.body !== undefined) {
      obj.body = ApplicationDetailRequestBody.toJSON(message.body);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ApplicationDetailRequest>, I>>(base?: I): ApplicationDetailRequest {
    return ApplicationDetailRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ApplicationDetailRequest>, I>>(object: I): ApplicationDetailRequest {
    const message = createBaseApplicationDetailRequest();
    message.header = (object.header !== undefined && object.header !== null)
      ? MessageHeader.fromPartial(object.header)
      : undefined;
    message.body = (object.body !== undefined && object.body !== null)
      ? ApplicationDetailRequestBody.fromPartial(object.body)
      : undefined;
    return message;
  },
};

function createBaseApplicationDetailRequestBody(): ApplicationDetailRequestBody {
  return { did: "", version: 0 };
}

export const ApplicationDetailRequestBody: MessageFns<ApplicationDetailRequestBody> = {
  encode(message: ApplicationDetailRequestBody, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.did !== "") {
      writer.uint32(10).string(message.did);
    }
    if (message.version !== 0) {
      writer.uint32(16).uint32(message.version);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ApplicationDetailRequestBody {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseApplicationDetailRequestBody();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.did = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.version = reader.uint32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ApplicationDetailRequestBody {
    return {
      did: isSet(object.did) ? globalThis.String(object.did) : "",
      version: isSet(object.version) ? globalThis.Number(object.version) : 0,
    };
  },

  toJSON(message: ApplicationDetailRequestBody): unknown {
    const obj: any = {};
    if (message.did !== "") {
      obj.did = message.did;
    }
    if (message.version !== 0) {
      obj.version = Math.round(message.version);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ApplicationDetailRequestBody>, I>>(base?: I): ApplicationDetailRequestBody {
    return ApplicationDetailRequestBody.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ApplicationDetailRequestBody>, I>>(object: I): ApplicationDetailRequestBody {
    const message = createBaseApplicationDetailRequestBody();
    message.did = object.did ?? "";
    message.version = object.version ?? 0;
    return message;
  },
};

function createBaseApplicationDetailResponse(): ApplicationDetailResponse {
  return { header: undefined, body: undefined };
}

export const ApplicationDetailResponse: MessageFns<ApplicationDetailResponse> = {
  encode(message: ApplicationDetailResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.header !== undefined) {
      MessageHeader.encode(message.header, writer.uint32(10).fork()).join();
    }
    if (message.body !== undefined) {
      ApplicationDetailResponseBody.encode(message.body, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ApplicationDetailResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseApplicationDetailResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.header = MessageHeader.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.body = ApplicationDetailResponseBody.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ApplicationDetailResponse {
    return {
      header: isSet(object.header) ? MessageHeader.fromJSON(object.header) : undefined,
      body: isSet(object.body) ? ApplicationDetailResponseBody.fromJSON(object.body) : undefined,
    };
  },

  toJSON(message: ApplicationDetailResponse): unknown {
    const obj: any = {};
    if (message.header !== undefined) {
      obj.header = MessageHeader.toJSON(message.header);
    }
    if (message.body !== undefined) {
      obj.body = ApplicationDetailResponseBody.toJSON(message.body);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ApplicationDetailResponse>, I>>(base?: I): ApplicationDetailResponse {
    return ApplicationDetailResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ApplicationDetailResponse>, I>>(object: I): ApplicationDetailResponse {
    const message = createBaseApplicationDetailResponse();
    message.header = (object.header !== undefined && object.header !== null)
      ? MessageHeader.fromPartial(object.header)
      : undefined;
    message.body = (object.body !== undefined && object.body !== null)
      ? ApplicationDetailResponseBody.fromPartial(object.body)
      : undefined;
    return message;
  },
};
// ...
export interface ApplicationServer extends UntypedServiceImplementation {
  /** 搜索应用 */
  search: handleUnaryCall<SearchApplicationRequest, SearchApplicationResponse>;
  /** 创建应用 */
  create: handleUnaryCall<CreateApplicationRequest, CreateApplicationResponse>;
  /** 应用详情 */
  detail: handleUnaryCall<ApplicationDetailRequest, ApplicationDetailResponse>;
  /** 删除应用 */
  delete: handleUnaryCall<DeleteApplicationRequest, DeleteApplicationResponse>;
}