// src/yeying/api/common/code.ts
export enum ResponseCodeEnum {
  /** RESPONSE_CODE_UNKNOWN - 未知编码，是一种占位符 */
  RESPONSE_CODE_UNKNOWN = 0,
  /**
   * OK - 含义：操作完成
   * 使用场景：成功创建、成功提交
   */
  OK = 1,
  /**
   * INVALID_ARGUMENT - 含义：提供的参数无效或不符合预期
   * 使用场景：当接收到客户端请求参数不正确时返回
   */
  INVALID_ARGUMENT = 2,
  /**
   * UNAUTHENTICATED - 含义：请求未通过身份验证
   * 使用场景：当请求需要身份验证但未提供有效凭据时返回
   */
  UNAUTHENTICATED = 3,
  /**
   * PERMISSION_DENIED - 含义：请求无法被执行，因为没有足够的权限
   * 使用场景：当客户端没有执行请求所需操作的权限时返回
   */
  PERMISSION_DENIED = 4,
  /**
   * NOT_FOUND - 含义：请求对象或资源未找到
   * 使用场景：当请求的对象或资源不存在时返回
   */
  NOT_FOUND = 5,
  /**
   * ALREADY_EXISTS - 含义：请求创建的对象已经存在
   * 使用场景：当请求创建的对象（如用户、资源等）已经存在时返回
   */
  ALREADY_EXISTS = 6,
  /**
   * LIMIT_EXCEEDED - 含义：超过了资源使用限制
   * 使用场景：当请求操作超过了预先设定的限制时返回
   */
  LIMIT_EXCEEDED = 7,
  /**
   * UNAVAILABLE - 含义：请求的服务当前不可用
   * 使用场景：当请求的服务不可用（例如服务器维护、过载等）时返回
   */
  UNAVAILABLE = 8,
  /**
   * UNKNOWN_ERROR - 含义：未知错误
   * 使用场景：当出现未知错误时返回或无法精确匹配到其他错误码时使用
   */
  UNKNOWN_ERROR = 9,
  /**
   * NETWORK_ERROR - 含义：网络错误
   * 使用场景：应为网络错误，无法发送到服务器
   */
  NETWORK_ERROR = 10,
  /**
   * INVALID_CERT - 含义：无效证书
   * 使用场景：浏览器无法识别证书，在浏览器中添加为受信任证书
   */
  INVALID_CERT = 11,
  /**
   * NOT_SUPPORTED - 含义：不支持
   * 使用场景：接口、服务不支持
   */
  NOT_SUPPORTED = 12,
  /**
   * DATA_CORRUPTED - 含义：数据损坏
   * 使用场景：验证数据的签名不通过，当和存储供应商签约资产时，用户发送到供应商的数据验证签名不通过
   */
  DATA_CORRUPTED = 13,
  UNRECOGNIZED = -1,
}